requirements.txt
dependency-injector
msedge-selenium-tools==3.141
configparser
openpyxl
pytest
datetime